添加登录拦截器和点赞 -- 点赞还未完成，无法获取登录的用户

+++ src/main/java/com/wulizhou/pets/model/entity/Pets.java
@@ -15,7 +15,7 @@ public class Pets {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "petId")
-	private Long petId;
+	private Integer petId;
 
 	/**
 	 * 宠物类型
